[FIX] 중복된 형 선언 제거

+++ src/main/java/com/sopterm/makeawish/common/Util.java
@@ -33,6 +33,6 @@ public class Util {
 	}
 
 	public static String calculateContribution(int price, int targetPrice) {
-		return String.format("%.0f", (double) price / (double) targetPrice * 100);
+		return String.format("%.0f", (double) price / targetPrice * 100);
 	}
 }
